css(feature/ujin/#5): boxes에props shadow(value) 추가, 폴더 이름 outerbox에서 boxes로 변경
