+++ yeoreum/src/pages/myPage/components/MyPage.tsx
@@ -1,8 +1,6 @@
-import React from 'react'
-
 function MyPage() {
   return (
-    <div>mypage</div>
+    <div></div>
   )
 }
